requirements.txt
keyring
PyCrypto
six
